+++ changelog.md
@@ -6,6 +6,14 @@
 ### What's Changed
 
 _(most recent changes are listed on top):_
+- Added config: due date days. With this option on duplication the due date will add these days. Or leave blank to remove the due date on duplication.
+
+
+## v1.1
+
+### What's Changed
+
+_(most recent changes are listed on top):_
 - Added config: prefix
 - Added config+feature: week pattern modification on duplication (needs the [WeekHelper](https://github.com/Tagirijus/WeekHelper) >=v2.7.0 plugin!)